Implement workaround for lack of DEFAULT insert.
Sqlite doesn't support `DEFAULT` as a value when inserting.  Ecto makes
use of this with other databases to insert multiple columns but use the
defaults for some columns in some rows.

This implements a workaround: instead of running a single sql statement
we run one per row, omitting any columns that aren't provided for each
individual row.  This causes the default values to be inserted for those
rows, giving us the same result as other databases.